Numerous changes.
- Added install target to makefiles.
- Makefiles build erpc_c as liberpc.a.
- Updated build and install instructions in README.md.
- Fixed some bugs in C/C++ infra.
- Removed RPMsg-Lite transport dependency on MCMgr.
- Renamed a couple readme.txt files to readme.md.
- Fixed some bugs with error checking code generation in erpcgen.
- Fixed issue with erpcgen tests creating symlinks on Windows when
  running under Python 2.7.
- Updated erpcgen Visual Studio project to use prebuilt boost 1.61.0.

+++ erpc_c/infra/static_queue.h
@@ -93,7 +93,7 @@ protected:
     {
         if ((m_head + 1) % m_capacity != m_tail)
         {
-            memcpy((void *)((uint32_t)m_data + m_head * m_elementSize), element, m_elementSize);
+            memcpy((void *)((uintptr_t)m_data + m_head * m_elementSize), element, m_elementSize);
             m_head = (m_head + 1) % m_capacity;
             return true;
         }
@@ -110,7 +110,7 @@ protected:
         void *element = NULL;
         if (m_tail != m_head)
         {
-            element = (void *)((uint32_t)m_data + m_tail * m_elementSize);
+            element = (void *)((uintptr_t)m_data + m_tail * m_elementSize);
             m_tail = (m_tail + 1) % m_capacity;
         }
         return element;
